requirements.txt
torch
torchvision
numpy
pandas
tqdm
matplotlib
seaborn
cython
pynput
opencv-python==4.5.3.56 
pyautogui
scipy
PyYAML
nvidia-pyindex
protobuf<4.21.3
onnxruntime-gpu
onnx>=1.9.0
onnx-simplifier>=0.3.6
coremltools
mss
rich
simple_pid
cryptography
requests
pyinstaller
tinyaes
tensorrt_python\whl\tensorrt-8.5.1.7-cp37-none-win_amd64.whl
onnx_graphsurgeon